src/pocApi/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
pillow
numpy
